Rename json->nlohmann for consistency with standard use of 3rd party lib (#251)
