+++ examples/demo-app/demo_app.cpp
@@ -24,7 +24,7 @@
 
 #include "args/args.hxx"
 #include "happly.h"
-#include "json/json.hpp"
+#include "nlohmann/json.hpp"
 
 #include "simple_dot_mesh_parser.h"
 #include "surface_mesh_io.h"
