+++ src/polyscope.cpp
@@ -16,7 +16,7 @@
 
 #include "stb_image.h"
 
-#include "json/json.hpp"
+#include "nlohmann/json.hpp"
 using json = nlohmann::json;
 
 #include "backends/imgui_impl_opengl3.h"
